[#1602] Add a switch for whether the ZIPs should be deleted after importing,r=Oliver
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/realty/trunk@8866 735d13b6-9817-0410-8766-e36946ffe9aa

+++ lib/class.tx_realty_openImmoImport.php
@@ -1023,10 +1023,16 @@ class tx_realty_openImmoImport {
 		}
 
 		$removedFiles = array();
+		$deleteImportedZips = $this->globalConfiguration->getConfigurationValueBoolean(
+			'deleteZipsAfterImport'
+		);
+
 		foreach ($this->getPathsOfZipsToExtract($importDirectory) as $currentPath) {
-			$removedZipArchive = $this->deleteFile($currentPath);
-			if ($removedZipArchive != '') {
-				$removedFiles[] = $removedZipArchive;
+			if ($deleteImportedZips) {
+				$removedZipArchive = $this->deleteFile($currentPath);
+				if ($removedZipArchive != '') {
+					$removedFiles[] = $removedZipArchive;
+				}
 			}
 			$this->deleteFile($this->getNameForExtractionFolder($currentPath));
 		}
